requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-Login==0.6.3
python-socketio==5.10.0
openai==1.51.2
watchdog==3.0.0
reportlab==4.0.7
python-dotenv==1.0.0
werkzeug==3.0.1
eventlet==0.33.3
gunicorn==21.2.0
markdown==3.5.1
schedule==1.2.0
